+++ ZenUnitTests/TestRunners/SpecificTestClassRunnerTests.cpp
@@ -214,7 +214,7 @@ namespace ZenUnit
    {
       ZenUnitArgs zenUnitArgs;
       zenUnitArgs.random = random;
-      zenUnitArgs.randomseed = Random<unsigned short>();
+      zenUnitArgs.randomseed = Random<unsigned>();
       GetArgs_ZenMock.Return(zenUnitArgs);
       if (expectRandomForEach)
       {
